+++ nx-plugin/src/generators/create-update/steps/feature-module.step.ts
@@ -18,19 +18,23 @@ export class FeatureModuleStep
       'declarations: [',
       `declarations: [${className}CreateUpdateComponent,`
     );
-    moduleContent = moduleContent.replace(
-      'declarations:',
-      `
-      providers: [providePortalDialogService()],
-      declarations:`
-    );
+
+    if (!moduleContent.includes('providePortalDialogService()')) {
+      moduleContent = moduleContent.replace(
+        'declarations:',
+        `
+    providers: [providePortalDialogService()],
+    declarations:`
+      );
+    }
+    
     moduleContent = moduleContent.replace(
       `from '@ngrx/effects';`,
       `from '@ngrx/effects';  
-       import { ${className}CreateUpdateComponent } from './pages/${fileName}-search/dialogs/${fileName}-create-update/${fileName}-create-update.component';
-       import { providePortalDialogService } from '@onecx/portal-integration-angular';`
+     import { ${className}CreateUpdateComponent } from './pages/${fileName}-search/dialogs/${fileName}-create-update/${fileName}-create-update.component';
+     import { providePortalDialogService } from '@onecx/portal-integration-angular';`
     );
-  
+
     tree.write(moduleFilePath, moduleContent);
   }
   getName(): string {